requirements.txt
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
idna==3.4
Pillow==9.5.0
requests==2.28.2
soupsieve==2.4
tqdm==4.65.0
urllib3==1.26.15
